+++ spec/lib/karafka/pro/processing/filters/virtual_limiter_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:limiter) { described_class.new(manager, collapser) }
+
+  pending
+end